simplify types (remove ethers)
