+++ packages/hardhat-plugin/package.json
@@ -52,10 +52,11 @@
     "typescript": "4.7.4"
   },
   "peerDependencies": {
-    "hardhat": "2.9.9"
+    "hardhat": "2.9.9",
+    "ethers": "5.7.0"
   },
   "dependencies": {
-    "@nomiclabs/hardhat-ethers": "2.0.6",
+    "@nomiclabs/hardhat-ethers": "^2.1.1",
     "@unlock-protocol/contracts": "0.0.8",
     "eslint-plugin-jest": "26.6.0",
     "eslint-plugin-mocha": "10.1.0"
